feat: Added client library
A client library has been added to the activity collector.
The client library can be used by other applications but primarily will be
used by our nutrition application's API Gateway.

This is the initial state of the client, more functionality to request
different types of data w/different parameters is to come.