test: Added initial test for activity model
models_init_test.go was added in which the TestMain for the models
package will live. The TestMain will be responsible for initializing
our connections or any other test setups.
The first test that has been added lives in activity_test.go, in
this file we are just testing the method that returns the sample
activity struct. Future activity model tests will live here.